wgpu: try, fail and giveup on rect rendering with wgpu
I've been at this for days. It refuses to let me do instanced drawingu
for rectangles. I have determined that there is nothing wrong with my
code and that it just plain is not working.

If the rect/vertex data is duplicated 4 times and positioned based on
the vertex index, it works.
But as soon as I attempt to use instancing it does not. There is nothing
wrong with the instancing logic, The indices get there.
Using an instance buffer with the indices also does not work.

So this just is not going to work.

Also wgpu refuses to use Vulkan on my machine even though it supports it

So like with windowing I am just having to do it myself.

Vulkan it is. DirectX it will be. And Metal will come one day.
Just like the Win32 api and whatever MacOs uses for windowing.

I should just learn my lesson by now.